Moved InputParserImproved and OutputRawImproved into irc package
