+++ src/main/java/com/github/otbproject/otbproject/irc/IRCBot.java
@@ -10,7 +10,6 @@ import com.github.otbproject.otbproject.channels.Channel;
 import com.github.otbproject.otbproject.channels.ChannelNotFoundException;
 import com.github.otbproject.otbproject.database.DatabaseWrapper;
 import com.github.otbproject.otbproject.serviceapi.ApiRequest;
-import com.github.otbproject.otbproject.util.OutputRawImproved;
 import org.isomorphism.util.TokenBucket;
 import org.isomorphism.util.TokenBuckets;
 import org.pircbotx.Configuration;
